requirements.txt
asgiref==3.3.1
Django==3.0.5
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-taggit==1.3.0
Pillow==7.1.2
pytz==2021.1
sqlparse==0.4.1
